Начал наполнять фрукты и ягоды

+++ app/src/main/java/ru/plumsoftware/avocado/data/base/model/healthy/Kpfc.kt
@@ -2,7 +2,7 @@ package ru.plumsoftware.avocado.data.base.model.healthy
 
 data class Kpfc(
     val kals: Int,
-    val fats: Int,
-    val proteins: Int,
-    val carbohydrates: Int,
+    val fats: Double,
+    val proteins: Double,
+    val carbohydrates: Double,
 )
